+++ CHANGELOG.md
@@ -7,6 +7,23 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-pre.5] - 2021-11-16
+
+### Changed
+
+- Property names so that they adhere to PascalCase.
+
+### Added
+- Added `XROrigin` menu item to SceneInspector that creates and configures a basic `XROrigin` in the scene.
+
+### Fixed
+
+- Serialization issue where upgrading to [XROrigin](xref:Unity.XR.CoreUtils.XROrigin) from [XRRig](https://docs.unity3d.com/Packages/com.unity.xr.interaction.toolkit@1.0/api/UnityEngine.XR.Interaction.Toolkit.XRRig.html) may cause references to be broken in the component.
+- Issue where [XROrigin.CameraInOriginSpacePos](xref:Unity.XR.CoreUtils.XROrigin.CameraInOriginSpacePos) was being miscalculated.
+- Issue where the custom Inspector for [XROrigin](xref:Unity.XR.CoreUtils.XROrigin) was not being used.
+- Warning message referencing an old property name when a Camera could not be found.
+- A reflection issue with [ScriptableSettingsBase.GetInstanceByType](xref:Unity.XR.CoreUtils.ScriptableSettingsBase.GetInstanceByType(System.Type)) by renaming `EditorScriptableSettings<T>.instance` to `EditorScriptableSettings<T>.Instance`.
+
 ## [2.0.0-pre.3] - 2021-11-03
 
 This is the first release of *XR Core Utilities* package.
@@ -26,5 +43,4 @@ This is the first release of *XR Core Utilities* package.
 
 ### Fixed
 
-- Company name in `LICENSE.md` file from "Unity Technologies ApS" to "Unity Technologies". 
-
+- Company name in `LICENSE.md` file from "Unity Technologies ApS" to "Unity Technologies".